refactor: 타입스크립트 파일 위치 변경
