--- src/components/Avatar/type.ts
@@ -1,7 +1,0 @@
-export interface AvatarProps {
-  src?: string;
-  size?: 'mini' | 'small' | 'middle' | 'large';
-  alt?: string;
-  style?: React.CSSProperties;
-  onClick?: () => void;
-}
