--- src/components/Badge/type.ts
@@ -1,4 +1,0 @@
-export interface BadgeProps {
-  count?: number;
-  style?: React.CSSProperties;
-}
