--- src/components/Card/type.ts
@@ -1,8 +1,0 @@
-interface CardProps {
-  width: string;
-  height: string;
-  shadowType?: 'medium' | 'large';
-  style?: React.CSSProperties;
-}
-
-export default CardProps;
