--- src/components/Dropdown/type.ts
@@ -1,4 +1,0 @@
-export interface DropdownProps {
-  channelId: string;
-  setChannelId: (value: string) => void;
-}
