+++ src/components/Input/index.tsx
@@ -1,7 +1,24 @@
 ﻿import { forwardRef, Ref } from 'react';
 
-import { InputProps } from '@/components/Input/type';
 import { Wrapper, StyledInput } from '@/components/Input/style';
+
+export interface InputProps
+  extends React.InputHTMLAttributes<HTMLInputElement> {
+  underline?: boolean;
+  fontType?: 'h1' | 'h2' | 'h3' | 'h4' | 'body1' | 'body2' | 'body3';
+  bordertype?:
+    | 'enabled'
+    | 'hover'
+    | 'focus'
+    | 'active'
+    | 'filled'
+    | 'error'
+    | 'disabled';
+  justifyContent?: 'center' | 'flex-start';
+
+  eyePosition?: string;
+  wrapperWidth?: string;
+}
 
 const Input = forwardRef<HTMLInputElement, InputProps>(
   (
