--- src/components/OnlineBadge/type.ts
@@ -1,4 +1,0 @@
-export interface OnlineBadgePropsTypes {
-  isOnline: boolean;
-  isFollowing: boolean;
-}
