--- src/components/Pagination/type.ts
@@ -1,5 +1,0 @@
-export interface PaginationProps {
-  page: number;
-  totalPage: number;
-  onPageChange: (page: number) => void;
-}
