--- src/components/ProtectedRoute/type.ts
@@ -1,3 +1,0 @@
-export interface ProtectedRouteProps {
-  admin?: boolean;
-}
