--- src/components/Tooltip/type.ts
@@ -1,9 +1,0 @@
-export type TooltipDirection = 'top' | 'bottom' | 'left' | 'right';
-export type TooltipType = 'click' | 'focus' | 'hover';
-
-export interface TooltipProps {
-  message: string;
-  type?: TooltipType;
-  direction: TooltipDirection;
-  hasArrow: boolean;
-}
