+++ src/components/layout/Header/NotificationCardBell/index.tsx
@@ -23,7 +23,15 @@ import {
 import UserSnippet from '@/pages/Main/components/UserList/UserSnippet';
 import { Badge, ScrollBar, Text, Button } from '@/components';
 import UserGroup from '@/pages/Main/components/UserList/UserGroup';
-import { NotificationData } from '@/components/layout/Header/NotificationCardBell/type';
+import { type User } from '@/types/APIResponseTypes';
+
+export interface NotificationData {
+  _id: string;
+  text: string;
+  isFollowing: boolean;
+  author: User;
+  handleClick?: () => void | undefined;
+}
 
 const NotificationCardBell = () => {
   const token = localStorage.getItem('auth-token');
