--- src/components/layout/Header/NotificationCardBell/type.ts
@@ -1,9 +1,0 @@
-import { User } from '@/types/APIResponseTypes';
-
-export interface NotificationData {
-  _id: string;
-  text: string;
-  isFollowing: boolean;
-  author: User;
-  handleClick?: () => void | undefined;
-}
