--- src/pages/Main/components/PostList/PostCard/type.ts
@@ -1,9 +1,0 @@
-import { Comment, User } from '@/types/APIResponseTypes';
-
-export interface PostCardProps {
-  image?: string;
-  title: string;
-  author: User;
-  postId: string;
-  comments: Comment[];
-}
