--- src/pages/Main/components/UserList/UserGroup/type.ts
@@ -1,3 +1,0 @@
-export interface UserSnippetGroupProps {
-  title?: string;
-}
