--- src/pages/Main/components/UserList/type.ts
@@ -1,10 +1,0 @@
-export type RenderUserSnippets = (
-  users: {
-    image: string;
-    isOnline: boolean;
-    isFollowing: boolean;
-    fullName: string;
-    _id: string;
-  }[],
-  title: string,
-) => JSX.Element;
