--- src/pages/PostDetail/components/Comment/Item/type.ts
@@ -1,13 +1,0 @@
-export interface CommentItemProps {
-  authorName: string;
-  authorImage: string;
-  createdAt: string;
-  comment: string;
-  authorId: string;
-  commentId: string;
-}
-
-export interface CommentContent {
-  type: string;
-  content: string;
-}
