--- src/pages/PostDetail/components/LikeButton/type.ts
@@ -1,4 +1,0 @@
-export interface LikeButtonProps {
-  postId: string | undefined;
-  userId: string | undefined;
-}
