--- src/pages/PostUpdate/component/PostForm/type.ts
@@ -1,8 +1,0 @@
-export interface FormType {
-  title: string;
-  content: string;
-  voteTitle: string;
-  voteArray: string[];
-  channelId: string;
-  image: File | null;
-}
