--- src/pages/PostUpdate/component/UpdateForm/type.ts
@@ -1,18 +1,0 @@
-export interface FormType {
-  title: string;
-  content: string;
-  voteTitle: string;
-  voteArray: string[];
-  channelId: string;
-
-  image: File | null;
-  imageSrc: string;
-  imageToDeletePublicId: string;
-}
-
-export interface TitleType {
-  title: string;
-  content: string;
-  voteTitle: string;
-  voteArray: string[];
-}
