--- src/pages/PostUpdate/component/VoteBox/type.ts
@@ -1,8 +1,0 @@
-export interface FormProps {
-  values: {
-    voteTitle: string;
-    voteArray: string[];
-  };
-  setFieldValue: (field: string, values: string | string[]) => void;
-  isEditable: boolean;
-}
