--- src/pages/Profile/components/GrassTable/type.ts
@@ -1,1 +1,0 @@
-export type LightnessType = 100 | 200 | 300 | 400 | 500;
