+++ src/pages/Sign/components/Up/index.tsx
@@ -1,10 +1,10 @@
+import { SignProps } from '../..';
 import { FormEvent, useCallback, useState } from 'react';
 import { AxiosError } from 'axios';
 import PasswordInput from '@/components/Input/PasswordInput';
 import { Button, Input } from '@/components';
 import { Form, Guide, Register, Warning } from '@/pages/Sign/components/style';
 import axiosInstance from '@/utils/customAxios';
-import { SignProps } from '@/pages/Sign/components/type';
 
 const Up = ({ isLogin, setIsLogin }: SignProps) => {
   const [email, setEmail] = useState('');
