--- src/pages/Sign/components/type.ts
@@ -1,4 +1,0 @@
-export interface SignProps {
-  isLogin: boolean;
-  setIsLogin: (isLogin: boolean) => void;
-}
